projeto3/src/java/main/br/com/dlima/domain/Produto.java
package br.com.dlima.domain;

import java.util.Date;

public class Produto extends Entity {
	private String nome;
	private Integer quantidade;
	private Date dataCadastro;
	public String getNome() {
		return nome;
	}
	public void setNome(String nome) {
		this.nome = nome;
	}
	public Integer getQuantidade() {
		return quantidade;
	}
	public void setQuantidade(Integer quantidade) {
		this.quantidade = quantidade;
	}
	public Date getDataCadastro() {
		return dataCadastro;
	}
	public void setDataCadastro(Date dataCadastro) {
		this.dataCadastro = dataCadastro;
	}

}
